pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nomad-r-runner"
version = "0.1.0"
description = "Submit R scripts as Nomad batch jobs running in Docker containers."
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-3.0-or-later"
dependencies = [
    "click>=8.0",
    "docker>=7.0",
    "psutil>=5.9",
    "python-nomad>=2.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
    "pdoc>=14.0",
]

[project.scripts]
nomad-r-runner = "nomad_r_runner.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
